updated delete api

+++ express-rest-api/routes/getusers.js
@@ -34,41 +34,6 @@ router.get('/profile', passport.authenticate('jwt', { session: false }), async (
 });
 
 
-router.post('/register', async (req, res) => {
-  try {
-    const { name, username, age, gender, password } = req.body;
-    const existingUser = await User.findOne({ username });
-
-    if (existingUser) {
-      return res.status(400).json({ error: 'Username already exists' });
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const userLog = new UserLogin({
-      username,
-      password: hashedPassword,
-      role: 'regular'
-    });
-
-    await userLog.save();
-
-    const user = new User({
-      username,
-      name,
-      age,
-      gender
-    });
-
-    await user.save();
-
-    res.status(201).json({ message: 'User registered successfully' });
-  }
-  catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Server error' });
-  }
-});
-
 router.put('/update/:username', passport.authenticate('jwt', { session: false }), async (req, res) => {
   try {
       const requestedUser = req.query.user;
@@ -125,6 +90,11 @@ router.delete('/delete/:username', passport.authenticate('jwt', { session: false
       }
 
       const { username } = req.params;
+
+      const existingUser = await UserLogin.findOne({ username });
+      if (!existingUser) {
+          return res.status(404).json({ error: 'User not found' });
+      }
 
       await UserLogin.findOneAndDelete({ username });
